+++ zebra-test/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## [3.0.0] - 2026-03-12
+
+### Breaking Changes
+
+- `mock_service::MockService::expect_no_requests` now returns `()` instead of its previous return type.
+
 ## [2.0.0] - 2025-10-15
 
 Test utilities were streamlined and Orchard vector fixtures were reorganized as an effort to
@@ -14,7 +22,6 @@ use `orchard::bundle::BatchValidator` in the transaction verifier ([#9308](https
 
 - `MockService::expect_no_requests` now returns `()` (it previously returned a value).
 - Removed now-unused test data ([#9308](https://github.com/ZcashFoundation/zebra/pull/9308))
-
 
 ## [1.0.1] - 2025-08-07